Ensure DOWN subports are cleaned up
Due to a known Neutron issue it's possible
to have subports with DOWN status and upon
controller restart those subports with are
not taken into account when repopulating the
pools resulting in leftovers. This commit
ensures nodes_ports_cleanup thread detach the
Port from the Trunk and remove it from the
Trunk.

Change-Id: If106637c81a5566d7bea00561f5cf1b381f97f23