Add config option to enable Kubernetes Event creation.
Kuryr can (or rather: soon will be able to) leverage events for certain
actions where Kuryr is involved, like creating pod networking, services,
network policies and so on. This is a nice addition to the logging,
where operator can just use typical kubectl commands to get the idea in
what state are his resources.

This feature can be turned on by using configuration:

[kubernetes]
use_events = true

in kuryr.conf, although it might have an performance impact on
kubernetes cluster.

Change-Id: Ia25e1309157e49992119709fb546fd5baa2462a4