Remove event added in KuryrNetwork's on_finalize
This commit removes creation of RemoveNeutronResourcesFailed event when
there's a problem on KuryrNetwork deletion. The event can only happen
when the Namespace related to the KuryrNetwork is being removed, meaning
that we're not able to create any object there, including Events.
Moreover the situation Event warns about is fairly common, indicating
that the pods in the process of being deleted from the namespace and
that haven't finished yet.

Change-Id: Ibb321ecdd6acf876320cfbabd77580addbc6359f