Add TRACE logging
There is a LOG.trace method and a TRACE level in oslo_log

https://github.com/openstack/oslo.log/blob/master/oslo_log/handlers.py#L38

We can utilize this log level to hide non-relevant debug messages.

At the moment with debug enabled, the logs are overwhelmed by amount of messages.

This log level can be enabled only with:

[DEFAULT]
default_log_levels=kuryr_kubernetes=TRACE

Closes-Bug: 1953090
Change-Id: I083dbe5f5c7309d091c28763cc38cde166282121