Pools: Fix order of updated SGs
According to the comments in vif_pool.py, if there are no ports in the
pool with the requested SG set, we should update the SG on another port,
starting from the ones that were created soonest. I think this logic is
to make sure we grab the ports with most outdated SGs. Anyway that code
is currently broken because of two issues:

1. _last_update dict is always updated by replacing whole dict,
   basically meaning that it's only holding data for SG that got updated
   most recently.
2. There's a race condition where in _get_port_from_pool multiple
   threads can steal a port from themselves.

This commit solves #2 by switching to use OrderedDict to track which SG
is the one that was used most recently. This way we can just iterate the
OrderDict when choosing which port should get its SG updated and just
choose the next port if we get IndexError on pop(). This also solves #1
because _last_update is no longer used to decide which ports have the
most outdated SGs.

Change-Id: Ia3159ee007be865db404e2dcef688abe21592553