CNI: Improve logging of timeout errors
There are 2 places we can time out in CNI:
1. KuryrPort haven't got created for a pod we got ADD request for.
2. Ports we plugged for a pod aren't moving to ACTIVE state.

This commit improves logging in these two cases by making sure an
exception with a meaningful message is raised when such a timeout
occurs. The messages includes clues on where to look for root the cause
next - kuryr-controller in case of 1 and Neutron in case of 2.

I'm also disabling putting kuryr-cni into unhealthy state in these
cases. Those erorrs cannot be cleared by a restart of the pod.

Change-Id: I5b881b58fe7d6dfed66a7bb6e3473b5b7939854d