Remove log that has no sense
The removed log is on Pod handler's on_finalize(). If we discover that
associated KuryrPort has deletionTimestamp set, we try to trigger its
on_finalize by setting an annotation on it. In case of 404 we were
logging an error. The problem is that we don't know if KuryrPorts
on_finalize wasn't really running so 404 is not really a problem, it
indicates that it was actually being deleted. Logging error doesn't make
sense there, this is an indicator things work as expected.

The situation when KuryrPort is deleted before the Pod is pretty common
when a deletion of a namespace happens, hence these logs are common.

Change-Id: I34618c9d6ba5e5e89eef7688cdffede2be21b489