Remove another log without sense
In Endpoints' on_delete we're attempting to remove spec.endpointSlices
from the KuryrLoadBalancer object. In case of 422 Unprocessable Entity
error we're logging a warning. Such code is however mostly returned when
our patch cannot be applied, so in this particular case when spec is
already missing the endpointSlices field. This patch makes sure we don't
log anything in that case as it's definitely not a warning.

Change-Id: If4f3173bd3bdf72862c2f0f94ed016b68eaccb30