Added missing raise statement for not ready resource.
Recently, there was Events feature added to the KuryrPort/vif handlers,
where in one case the ResourceNotReady exception needed be to handled
elsewhere, but on behalf of event handling we needed to catch it on
KuryrPort handler, but it wasn't raised causing to swallow up the
events, and slowing down ports acknowledge. This patch fixes this.

Change-Id: I203773de2b8c0d1b21ef3f722df7ed7da562a892