Ignore Pod event if Pod is Pending with no Node set
It's possible that a Pod had a nodeName configured and
Kuryr started handling it, but after a while it was not
scheduled in that node anymore. This impacts the processing
of the Pod event as there are steps in between that the
Pod object is retrieved again with possibly no nodeName
assigned, making the controller to constantly retry for an
event that should be ignored.

Closes-bug: 1956767
Change-Id: I123dc706e14aa3edae06a165d6c0c09e99ef28d6