Update Centos version for container images.
Due to the EOL for Centos 8, we need to go forward, and adopt to the
Stream version of Centos for container images of kuryr CNI and
controller.

There was applied modification for fixing (hopefully transient) issues
with variable usage in yum/dnf, so there is a need for replacing those
manually.

And finally, there was a switch from deprecated yum to dnf.

Change-Id: I19c16877e8ba6f401c9d76ed70b2380c4e3cfbe0