CI: Cleanup what we gather from K8s
Seems like the set of resources we gather from K8s for CI gates
debugging purposes was outdated. This commit updates it to make sure
we're also getting NPs and KLBs, while removing Ingress, KuryrNet and
KuryrNetPolicy that are obsolete today.

Besides that KuryrNet and KuryrNetPolicy CRD definitions are removed as
we're not using them for a long time now.

Change-Id: Ia7ecd8f79b6ac0a8172cf108e228e2ea1cb7a0d5