Pin setuptools in CNI to 53.0.0-10.el9
Looks like RDO repos upgrade setuptools to version that creates a
problem when we attempt to upgrade it through pip. This commit pins
RPM setuptools to a version available in baseos repos.

This isn't really a great solution, we should probably just start
running in a virtualenv, but let's try to unblock the gates with this
thing.

Change-Id: I0b9e02ef1e9227497082a4a3cd803c52ad2789fb