Run Kuryr inside containers in venv
We've just hit an issue where Ceph repo added by RDO upgraded setuptools
to a version that pip cannot uninstall. This is mainly because we're
still installing all the Kuryr dependencies in system's site-packages.
This commit switches us to create and use a virtualenv to have a clean
environment in which we install all the dependencies through pip.

Change-Id: Ieb9fd5ed0251425e9fe172e4a93ad048768ce785