Lock pool port creation during prepopulation.
It might happen that during several pod creation there would be
triggered port population for the same namespace when using nested
driver. And since there is bulk port creation involved it might be
uneasy for Neutron. Solution for this would be locking port creation
operation, so that we make sure, there is only one population happening
at a time.

Closes-Bug: #1982478
Change-Id: I6e55963374b9c4105bf759c16add286def559491