Force kvm and host-passthrough for Amphora jobs
We face general instability of the Amphora jobs that we have. A while
ago we've made sure that these jobs are scheduled on the clouds that
support nested virtualization in order to boost performance of the
Amphora LB VMs. Now it turns out we haven't configured DevStack to
actually use kvm instead of qemu, so it's possible that we were still
running these jobs with software virtualization.

This commit attempts to reconfigure DevStack to actually use kvm and
nested virtualization.

Change-Id: I57f70c1cb140d4fa54d13b6c759a0179d8adb7b4
(cherry picked from commit d65720cf7434cd2e201415535d316d3189873d9a)
Depends-On: https://review.opendev.org/c/openstack/neutron/+/848131