Use either subnet name or id for Machines.
Currently, we support only subnet id for primarySubnet field for
OpenShift Machines. Even though it's possible to create objects in
OpenStack with the same name, it is more natural to use names instead of
id especially in OpenShift world. In this patch we introduce support
for using names as well.

Change-Id: Ib21646b4b7cf0e3c07ddef15b3569a0fb4539e8a