Make Amphora jobs non-voting
For a long time gate jobs with Amphora are unstable due to LBs being
unstable due to Amphora somehow not working as expected. The gates with
OVN Octavia provider are much more stable for the same cases. This
commit makes all the jobs with Amphora non-voting to fight constant
rechecks that we're forced to do.

Change-Id: I94a7403fa1dd038b7f9f8a922a4f635ed5816067