Skip retry of Network Policy event
When attempting to Handle a Network Policy and one Namespace
which is affected by it is in the process of getting handled
by Kuryr, the Network Policy event would be often retried.
This commit removes the retry to make sure the Network Policy
gets updated only once the Namespace handling has finsihed.

Change-Id: I73c9488dca21f73070ca84352e3ba3780ea7298f